requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
pydantic>=2.0.0
sse-starlette>=1.6.0
weasyprint>=60.0
